making common base class
