+++ ethsigner/core/src/main/java/tech/pegasys/ethsigner/core/requesthandler/passthrough/PassThroughHandler.java
@@ -12,15 +12,11 @@
  */
 package tech.pegasys.ethsigner.core.requesthandler.passthrough;
 
-import static io.netty.handler.codec.http.HttpResponseStatus.GATEWAY_TIMEOUT;
-import static io.netty.handler.codec.http.HttpResponseStatus.INTERNAL_SERVER_ERROR;
-
 import tech.pegasys.ethsigner.core.jsonrpc.JsonRpcRequest;
 import tech.pegasys.ethsigner.core.requesthandler.JsonRpcRequestHandler;
+import tech.pegasys.ethsigner.core.requesthandler.utils.ResponseHandler;
 
-import java.net.ConnectException;
 import java.time.Duration;
-import java.util.concurrent.TimeoutException;
 
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpClient;
@@ -32,7 +28,7 @@ import io.vertx.ext.web.RoutingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class PassThroughHandler implements JsonRpcRequestHandler {
+public class PassThroughHandler extends ResponseHandler implements JsonRpcRequestHandler {
 
   private static final Logger LOG = LogManager.getLogger();
 
@@ -55,65 +51,20 @@ public class PassThroughHandler implements JsonRpcRequestHandler {
             response -> handleResponse(context, response));
 
     proxyRequest.setTimeout(httpRequestTimeout.toMillis());
-    proxyRequest.exceptionHandler(thrown -> requestExceptionHandler(context, thrown));
+    proxyRequest.exceptionHandler(thrown -> handleException(context, thrown));
     proxyRequest.headers().setAll(httpServerRequest.headers());
     proxyRequest.setChunked(false);
-
     proxyRequest.end(context.getBody());
     logRequest(request, httpServerRequest);
   }
 
-  private void requestExceptionHandler(final RoutingContext context, final Throwable thrown) {
-    LOG.info("An exception was thrown by the transaction submission, {}", thrown);
-    if (thrown instanceof TimeoutException) {
-      context.fail(GATEWAY_TIMEOUT.code());
-    } else if (thrown instanceof ConnectException) {
-      context.fail(GATEWAY_TIMEOUT.code());
-    } else {
-      context.fail(INTERNAL_SERVER_ERROR.code());
-    }
-  }
-
-  private void handleResponse(final RoutingContext context, final HttpClientResponse response) {
-    logResponse(response);
-
-    response.bodyHandler(
-        body -> {
-          LOG.info("Executing Body Handler");
-          context
-              .vertx()
-              .executeBlocking(
-                  future -> {
-                    logResponseBody(body);
-                    handleResponseBody(context, response, body);
-                    future.complete();
-                  },
-                  false,
-                  (res) -> {
-                    if (res.failed()) {
-                      LOG.error(
-                          "An unhandled error occurred while processing {}",
-                          context.getBodyAsString(),
-                          res.cause());
-                    }
-                  });
-        });
-  }
-
-  private void handleResponseBody(
+  @Override
+  protected void handleResponseBody(
       final RoutingContext context, final HttpClientResponse response, final Buffer body) {
     context.request().response().setStatusCode(response.statusCode());
     context.request().response().headers().setAll(response.headers());
     context.request().response().setChunked(false);
     context.request().response().end(body);
-  }
-
-  private void logResponse(final HttpClientResponse response) {
-    LOG.debug("Response status: {}", response.statusCode());
-  }
-
-  private void logResponseBody(final Buffer body) {
-    LOG.debug("Response body: {}", body);
   }
 
   private void logRequest(final JsonRpcRequest jsonRequest, final HttpServerRequest httpRequest) {
